use text fragments to highlight cited sentences
The bucket union website says the anthem was composed by Sachinee, but
it is not easy for readers to spot the statement. Now the text fragments
mechanism comes to rescue. It allows us to highlight a specific potion
of text in an external website so readers can readily locate the cited
statement.